Struct util::deque::Stealer

pub struct Stealer<T> {
    // some fields omitted
}

The stealing half of the work-stealing deque. Stealers have access to the opposite end of the deque from the worker, and they only have access to the steal method.

Methods

impl<T: Send + 'static> Stealer<T>

fn steal(&self) -> Stolen<T>

Steals work off the end of the queue (opposite of the worker's end)

fn pool<'a>(&'a self) -> &'a BufferPool<T>

Gets access to the buffer pool that this stealer is attached to. This can be used to create more deques which share the same buffer pool as this deque.