Module util::fnv

This file stolen wholesale from rustc/src/librustc/util/nodemap.rs

Structs

FnvHasher
A speedy hash algorithm for node ids and def ids. The hashmap in libcollections by default uses SipHash which isn't quite as speedy as we want. In the compiler we're not really worried about DOS attempts, so we just default to a non-cryptographic hash.