Module util::geometry

Structs

Au

Enums

PagePx
One CSS "px" in the root coordinate system for the content document.

ScreenPx
A normalized "pixel" at the default resolution for the display.

ViewportPx
One CSS "px" in the coordinate system of the "initial viewport": http://www.w3.org/TR/css-device-adapt/#initial-viewport

Constants

MAX_AU

MIN_AU

Statics

MAX_RECT

ZERO_POINT

ZERO_RECT

Functions

f32_rect_to_au_rect
A helper function to convert a rect of f32 pixels to a rect of app units.

from_frac_px

from_pt

from_px

pt_to_px

px_to_pt

rect_contains_point
Returns true if the rect contains the given point. Points on the top or left sides of the rect are considered inside the rectangle, while points on the right or bottom sides of the rect are not considered inside the rectangle.

to_frac_px

to_pt

to_px