Enum util::geometry::ViewportPx

pub enum ViewportPx {}

One CSS "px" in the coordinate system of the "initial viewport": http://www.w3.org/TR/css-device-adapt/#initial-viewport

ViewportPx is equal to ScreenPx times a "page zoom" factor controlled by the user. This is the desktop-style "full page" zoom that enlarges content but then reflows the layout viewport so it still exactly fits the visible area.

At the default zoom level of 100%, one PagePx is equal to one ScreenPx. However, if the document is zoomed in or out then this scale may be larger or smaller.