Module util::memory

Memory profiling functions.

Structs

MemoryProfiler

MemoryProfilerChan

MemoryReport

MemoryReportsChan
A channel through which memory reports can be sent.

struct_mallinfo

Enums

MemoryProfilerMsg
Messages that can be sent to the memory profiler thread.

Traits

MemoryReporter
A memory reporter is capable of measuring some data structure of interest. Because it needs to be passed to and registered with the MemoryProfiler, it's typically a "small" (i.e. easily cloneable) value that provides access to a "large" data structure, e.g. a channel that can inject a request for measurements into the event queue associated with the "large" data structure.

SizeOf

Functions

heap_size_of