Module util::opts

Configuration options for a single run of the servo application. Created from command line arguments.

Structs

Opts
Global flags for Servo, currently set on the command line.

Functions

default_opts

experimental_enabled

from_cmdline_args

get

print_debug_usage

set_experimental_enabled

set_opts