Struct util::opts::DebugOptions

pub struct DebugOptions {
    pub help: bool,
    pub bubble_widths: bool,
    pub disable_text_aa: bool,
    pub disable_canvas_aa: bool,
    pub dump_flow_tree: bool,
    pub dump_display_list: bool,
    pub dump_display_list_json: bool,
    pub dump_layer_tree: bool,
    pub relayout_event: bool,
    pub profile_script_events: bool,
    pub profile_heartbeats: bool,
    pub show_compositor_borders: bool,
    pub show_fragment_borders: bool,
    pub show_parallel_paint: bool,
    pub show_parallel_layout: bool,
    pub paint_flashing: bool,
    pub trace_layout: bool,
    pub disable_share_style_cache: bool,
    pub convert_mouse_to_touch: bool,
    pub replace_surrogates: bool,
    pub gc_profile: bool,
    pub load_webfonts_synchronously: bool,
    pub disable_vsync: bool,
    pub webrender_stats: bool,
    pub use_msaa: bool,
    pub full_backtraces: bool,
}

Debug options for Servo, currently set on the command line with -Z

Fields

help

List all the debug options.

bubble_widths

Bubble intrinsic widths separately like other engines.

disable_text_aa

Disable antialiasing of rendered text.

disable_canvas_aa

Disable antialiasing of rendered text on the HTML canvas element.

dump_flow_tree

Print the flow tree after each layout.

dump_display_list

Print the display list after each layout.

dump_display_list_json

Print the display list in JSON form.

dump_layer_tree

Print the layer tree whenever it changes.

relayout_event

Print notifications when there is a relayout.

profile_script_events

Profile which events script threads spend their time on.

profile_heartbeats

Enable all heartbeats for profiling.

show_compositor_borders

Paint borders along layer and tile boundaries.

show_fragment_borders

Paint borders along fragment boundaries.

show_parallel_paint

Overlay tiles with colors showing which thread painted them.

show_parallel_layout

Mark which thread laid each flow out with colors.

paint_flashing

Overlay repainted areas with a random color.

trace_layout

Write layout trace to an external file for debugging.

disable_share_style_cache

Disable the style sharing cache.

convert_mouse_to_touch

Translate mouse input into touch events.

replace_surrogates

Replace unpaires surrogates in DOM strings with U+FFFD. See https://github.com/servo/servo/issues/6564

gc_profile

Log GC passes and their durations.

load_webfonts_synchronously

Load web fonts synchronously to avoid non-deterministic network-driven reflows.

disable_vsync

Disable vsync in the compositor

webrender_stats

Show webrender profiling stats on screen.

use_msaa

Use multisample antialiasing in WebRender.

full_backtraces

Methods

impl DebugOptions

fn new(debug_string: &str) -> Result<DebugOptions, &str>