Struct util::opts::Opts

pub struct Opts {
    pub urls: Vec<String>,
    pub paint_threads: usize,
    pub gpu_painting: bool,
    pub tile_size: usize,
    pub device_pixels_per_px: Option<ScaleFactor<ScreenPx, DevicePixel, f32>>,
    pub time_profiler_period: Option<f64>,
    pub mem_profiler_period: Option<f64>,
    pub enable_experimental: bool,
    pub layout_threads: usize,
    pub nonincremental_layout: bool,
    pub nossl: bool,
    pub userscripts: bool,
    pub output_file: Option<String>,
    pub headless: bool,
    pub hard_fail: bool,
    pub bubble_inline_sizes_separately: bool,
    pub show_debug_borders: bool,
    pub show_debug_fragment_borders: bool,
    pub show_debug_parallel_paint: bool,
    pub show_debug_parallel_layout: bool,
    pub enable_text_antialiasing: bool,
    pub trace_layout: bool,
    pub profile_tasks: bool,
    pub devtools_port: Option<u16>,
    pub initial_window_size: TypedSize2D<ScreenPx, u32>,
    pub user_agent: Option<String>,
    pub dump_flow_tree: bool,
    pub dump_display_list: bool,
    pub relayout_event: bool,
    pub validate_display_list_geometry: bool,
    pub resources_path: Option<String>,
}

Global flags for Servo, currently set on the command line.

Fields

urls

The initial URLs to load.

paint_threads

How many threads to use for CPU painting (-t).

Note that painting is sequentialized when using GPU painting.

gpu_painting

True to use GPU painting via Skia-GL, false to use CPU painting via Skia (-g). Note that compositing is always done on the GPU.

tile_size

The maximum size of each tile in pixels (-s).

device_pixels_per_px

The ratio of device pixels per px at the default scale. If unspecified, will use the platform default setting.

time_profiler_period

None to disable the time profiler or Some with an interval in seconds to enable it and cause it to produce output on that interval (-p).

mem_profiler_period

None to disable the memory profiler or Some with an interval in seconds to enable it and cause it to produce output on that interval (-m).

enable_experimental

Enable experimental web features (-e).

layout_threads

The number of threads to use for layout (-y). Defaults to 1, which results in a recursive sequential algorithm.

nonincremental_layout

nossl

userscripts

output_file

headless

hard_fail

bubble_inline_sizes_separately

True if we should bubble intrinsic widths sequentially (-b). If this is true, then intrinsic widths are computed as a separate pass instead of during flow construction. You may wish to turn this flag on in order to benchmark style recalculation against other browser engines.

show_debug_borders

True if we should show borders on all layers and tiles for debugging purposes (--show-debug-borders).

show_debug_fragment_borders

True if we should show borders on all fragments for debugging purposes (--show-debug-fragment-borders).

show_debug_parallel_paint

True if we should paint tiles with overlays based on which thread painted them.

show_debug_parallel_layout

True if we should paint borders around flows based on which thread painted them.

enable_text_antialiasing

If set with --disable-text-aa, disable antialiasing on fonts. This is primarily useful for reftests where pixel perfect results are required when using fonts such as the Ahem font for layout tests.

trace_layout

True if each step of layout is traced to an external JSON file for debugging purposes. Settings this implies sequential layout and paint.

profile_tasks

If true, instrument the runtime for each task created and dump that information to a JSON file that can be viewed in the task profile viewer.

devtools_port

None to disable devtools or Some with a port number to start a server to listen to remote Firefox devtools connections.

initial_window_size

The initial requested size of the window.

user_agent

An optional string allowing the user agent to be set for testing.

dump_flow_tree

Dumps the flow tree after a layout.

dump_display_list

Dumps the flow tree after a layout.

relayout_event

Emits notifications when there is a relayout.

validate_display_list_geometry

Whether to show an error when display list geometry escapes flow overflow regions.

resources_path

A specific path to find required resources (such as user-agent.css).