Module util::remutex

An implementation of re-entrant mutexes.

Re-entrant mutexes are like mutexes, but where it is expected that a single thread may own a lock more than once. It provides the same interface as https://github.com/rust-lang/rust/blob/master/src/libstd/sys/common/remutex.rs so if those types are ever exported, we should be able to replace this implemtation.

Structs

AtomicOptThreadId
A type for atomic storage of thread ids.

HandOverHandMutex
A type for hand-over-hand mutexes.

ReentrantMutex
A type for re-entrant mutexes.

ReentrantMutexGuard

THREAD_COUNT

ThreadId
A type for thread ids.

Statics

THREAD_COUNT

THREAD_ID