Struct util::selectors::Selector

pub struct Selector<T, R: Rng> { /* fields omitted */ }

Selector struct

Methods

impl<T> Selector<T, ThreadRng>

fn new_default_rng() -> Self

Create a new selector using a default random number generator.

Equivalent to Selector::new(rand::thread_rng()).

impl<T, R: Rng> Selector<T, R>

fn new(rng: R) -> Self

Create a new selector

fn from_iterator<I, F>(it: I, oddf: F, rng: R) -> Selector<T, R> where I: IntoIterator<Item=T>, F: Fn(&T) -> f64

Consume the iterator, creating a new selector

fn update(&mut self, odds: f64, value: T) -> Option<T>

Update the selector, replacing the original value with a new value with a probability p/sum(p)

fn decompose(self) -> Option<T>

Consumes itself, yielding a random value, or nothing if the selector had no data.