Trait util::smallvec::SmallVec

pub trait SmallVec<T> where Self: SmallVecPrivate<T> {
    fn inline_size(&self) -> usize;
    fn len(&self) -> usize;
    fn is_empty(&self) -> bool;
    fn cap(&self) -> usize;

    fn spilled(&self) -> bool { ... }
    fn begin(&self) -> *const T { ... }
    fn begin_mut(&mut self) -> *mut T { ... }
    fn end(&self) -> *const T { ... }
    fn end_mut(&mut self) -> *mut T { ... }
    fn iter(&'a self) -> SmallVecIterator<'a, T> { ... }
    fn mut_iter(&'a mut self) -> SmallVecMutIterator<'a, T> { ... }
    fn into_iter(&'a mut self) -> SmallVecMoveIterator<'a, T> { ... }
    fn push(&mut self, value: T) { ... }
    fn push_all_move<V>(&mut self, other: V) where V: SmallVec<T> { ... }
    fn pop(&mut self) -> Option<T> { ... }
    fn grow(&mut self, new_cap: usize) { ... }
    fn get(&'a self, index: usize) -> &'a T { ... }
    fn get_mut(&'a mut self, index: usize) -> &'a mut T { ... }
    fn slice(&'a self, start: usize, end: usize) -> &'a [T] { ... }
    fn as_slice(&'a self) -> &'a [T] { ... }
    fn as_slice_mut(&'a mut self) -> &'a mut [T] { ... }
    fn slice_mut(&'a mut self, start: usize, end: usize) -> &'a mut [T] { ... }
    fn slice_from_mut(&'a mut self, start: usize) -> &'a mut [T] { ... }
    fn fail_bounds_check(&self, index: usize) { ... }
}

Required Methods

fn inline_size(&self) -> usize

fn len(&self) -> usize

fn is_empty(&self) -> bool

fn cap(&self) -> usize

Provided Methods

fn spilled(&self) -> bool

fn begin(&self) -> *const T

fn begin_mut(&mut self) -> *mut T

fn end(&self) -> *const T

fn end_mut(&mut self) -> *mut T

fn iter(&'a self) -> SmallVecIterator<'a, T>

fn mut_iter(&'a mut self) -> SmallVecMutIterator<'a, T>

fn into_iter(&'a mut self) -> SmallVecMoveIterator<'a, T>

NB: For efficiency reasons (avoiding making a second copy of the inline elements), this actually clears out the original array instead of moving it.

fn push(&mut self, value: T)

fn push_all_move<V>(&mut self, other: V) where V: SmallVec<T>

fn pop(&mut self) -> Option<T>

fn grow(&mut self, new_cap: usize)

fn get(&'a self, index: usize) -> &'a T

fn get_mut(&'a mut self, index: usize) -> &'a mut T

fn slice(&'a self, start: usize, end: usize) -> &'a [T]

fn as_slice(&'a self) -> &'a [T]

fn as_slice_mut(&'a mut self) -> &'a mut [T]

fn slice_mut(&'a mut self, start: usize, end: usize) -> &'a mut [T]

fn slice_from_mut(&'a mut self, start: usize) -> &'a mut [T]

fn fail_bounds_check(&self, index: usize)