Module util::task_state

Supports dynamic assertions in debug builds about what sort of task is running and what state it's in.

In release builds, get returns 0. All of the other functions inline away to nothing.

Structs

TaskState

Constants

IN_GC

IN_HTML_PARSER

IN_WORKER

LAYOUT

PAINT

SCRIPT

Functions

enter

exit

get

initialize