Module util::thread_state

Supports dynamic assertions in debug builds about what sort of thread is running and what state it's in.

In release builds, get returns 0. All of the other functions inline away to nothing.

Modules

imp

Structs

ThreadState

Constants

IN_GC

IN_HTML_PARSER

IN_WORKER

LAYOUT

PAINT

SCRIPT

Statics

TYPES

Functions

enter

exit

get

initialize