Struct util::workqueue::WorkQueue

pub struct WorkQueue<QueueData: 'static, WorkData: 'static> {
    pub data: QueueData,
    // some fields omitted
}

A work queue on which units of work can be submitted.

Fields

data

Arbitrary user data.

Methods

impl<QueueData: Send, WorkData: Send> WorkQueue<QueueData, WorkData>

fn new(task_name: &'static str, state: TaskState, thread_count: usize, user_data: QueueData) -> WorkQueue<QueueData, WorkData>

Creates a new work queue and spawns all the threads associated with it.

fn push(&mut self, work_unit: WorkUnit<QueueData, WorkData>)

Enqueues a block into the work queue.

fn run(&mut self)

Synchronously runs all the enqueued tasks and waits for them to complete.

fn shutdown(&mut self)