Struct util::workqueue::WorkerThread

struct WorkerThread<QueueData: 'static, WorkData: 'static + Send> {
    index: usize,
    port: Receiver<WorkerMsg<QueueData, WorkData>>,
    chan: Sender<SupervisorMsg<QueueData, WorkData>>,
    other_deques: Vec<Stealer<WorkUnit<QueueData, WorkData>>>,
    rng: XorShiftRng,
}

Information specific to each worker thread that the thread keeps.

Fields

index

The index of this worker.

port

The communication port from the supervisor.

chan

The communication channel on which messages are sent to the supervisor.

other_deques

The thief end of the work-stealing deque for all other workers.

rng

The random number generator for this worker.

Methods

impl<QueueData: Sync, WorkData: Send> WorkerThread<QueueData, WorkData>

fn start(&mut self)

The main logic. This function starts up the worker and listens for messages.