Crate util

Modules

bezier
Parametric Bézier curves.

cache

cursor
A list of common mouse cursors per CSS3-UI  8.1.1.

debug_utils

deque
A (mostly) lock-free concurrent work-stealing deque

fnv
This file stolen wholesale from rustc/src/librustc/util/nodemap.rs

geometry

linked_list
Utility functions for doubly-linked lists.

logical_geometry
Geometry in flow-relative space.

mem
Data structure measurement.

namespace

opts
Configuration options for a single run of the servo application. Created from command line arguments.

persistent_list
A persistent, thread-safe singly-linked list.

range

resource_files

smallvec
Small vectors in various sizes. These store a certain number of elements inline and fall back to the heap for larger allocations.

str

task

task_state
Supports dynamic assertions in debug builds about what sort of task is running and what state it's in.

taskpool
A load-balancing task pool.

tid

vec

workqueue
A work queue for scheduling units of work across threads in a fork-join fashion.

Macros

int_range_index!
Implements a range index type with operator overloads

Functions

arc_ptr_eq

breakpoint