Module util::opts

Configuration options for a single run of the servo application. Created from command line arguments.

Structs

DebugOptions
Debug options for Servo, currently set on the command line with -Z

OPTIONS

Opts
Global flags for Servo, currently set on the command line.

Enums

ArgumentParsingResult

OutputOptions

RenderApi

UserAgent

Constants

DEFAULT_RENDER_API

DEFAULT_USER_AGENT

INVALID_OPTIONS

_IMPL_DESERIALIZE_FOR_Opts

_IMPL_DESERIALIZE_FOR_OutputOptions

_IMPL_DESERIALIZE_FOR_RenderApi

_IMPL_SERIALIZE_FOR_Opts

_IMPL_SERIALIZE_FOR_OutputOptions

_IMPL_SERIALIZE_FOR_RenderApi

Statics

DEFAULT_OPTIONS

EXPERIMENTAL_ENABLED

FORCE_CPU_PAINTING

MULTIPROCESS

OPTIONS

Functions

args_fail

default_opts

default_user_agent_string

from_cmdline_args

get

multiprocess

parse_url_or_filename

print_debug_usage

print_usage

set_defaults

set_experimental_enabled
Turn on experimental features globally. Normally this is done during initialization by set or from_cmdline_args, but tests that require experimental features will also set it.