Struct util::opts::Opts

pub struct Opts {
    pub is_running_problem_test: bool,
    pub url: Option<Url>,
    pub paint_threads: usize,
    pub gpu_painting: bool,
    pub tile_size: usize,
    pub device_pixels_per_px: Option<f32>,
    pub time_profiling: Option<OutputOptions>,
    pub time_profiler_trace_path: Option<String>,
    pub mem_profiler_period: Option<f64>,
    pub layout_threads: usize,
    pub nonincremental_layout: bool,
    pub userscripts: Option<String>,
    pub user_stylesheets: Vec<(Vec<u8>, Url)>,
    pub output_file: Option<String>,
    pub replace_surrogates: bool,
    pub gc_profile: bool,
    pub load_webfonts_synchronously: bool,
    pub headless: bool,
    pub hard_fail: bool,
    pub bubble_inline_sizes_separately: bool,
    pub show_debug_borders: bool,
    pub show_debug_fragment_borders: bool,
    pub show_debug_parallel_paint: bool,
    pub show_debug_parallel_layout: bool,
    pub paint_flashing: bool,
    pub enable_text_antialiasing: bool,
    pub enable_canvas_antialiasing: bool,
    pub trace_layout: bool,
    pub profile_script_events: bool,
    pub profile_heartbeats: bool,
    pub devtools_port: Option<u16>,
    pub webdriver_port: Option<u16>,
    pub initial_window_size: TypedSize2D<ScreenPx, u32>,
    pub user_agent: String,
    pub multiprocess: bool,
    pub sandbox: bool,
    pub random_pipeline_closure_probability: Option<f32>,
    pub random_pipeline_closure_seed: Option<usize>,
    pub dump_flow_tree: bool,
    pub dump_display_list: bool,
    pub dump_display_list_json: bool,
    pub dump_layer_tree: bool,
    pub relayout_event: bool,
    pub disable_share_style_cache: bool,
    pub convert_mouse_to_touch: bool,
    pub exit_after_load: bool,
    pub no_native_titlebar: bool,
    pub enable_vsync: bool,
    pub use_webrender: bool,
    pub webrender_stats: bool,
    pub use_msaa: bool,
    pub profile_dir: Option<String>,
    pub render_api: RenderApi,
    pub full_backtraces: bool,
}

Global flags for Servo, currently set on the command line.

Fields

is_running_problem_test

url

The initial URL to load.

paint_threads

How many threads to use for CPU painting (-t).

Note that painting is sequentialized when using GPU painting.

gpu_painting

True to use GPU painting via Skia-GL, false to use CPU painting via Skia (-g). Note that compositing is always done on the GPU.

tile_size

The maximum size of each tile in pixels (-s).

device_pixels_per_px

The ratio of device pixels per px at the default scale. If unspecified, will use the platform default setting.

time_profiling

None to disable the time profiler or Some with an interval in seconds to enable it and cause it to produce output on that interval (-p).

time_profiler_trace_path

When the profiler is enabled, this is an optional path to dump a self-contained HTML file visualizing the traces as a timeline.

mem_profiler_period

None to disable the memory profiler or Some with an interval in seconds to enable it and cause it to produce output on that interval (-m).

layout_threads

The number of threads to use for layout (-y). Defaults to 1, which results in a recursive sequential algorithm.

nonincremental_layout

userscripts

Where to load userscripts from, if any. An empty string will load from the resources/user-agent-js directory, and if the option isn't passed userscripts won't be loaded

user_stylesheets

output_file

replace_surrogates

Replace unpaires surrogates in DOM strings with U+FFFD. See https://github.com/servo/servo/issues/6564

gc_profile

Log GC passes and their durations.

load_webfonts_synchronously

Load web fonts synchronously to avoid non-deterministic network-driven reflows.

headless

hard_fail

bubble_inline_sizes_separately

True if we should bubble intrinsic widths sequentially (-b). If this is true, then intrinsic widths are computed as a separate pass instead of during flow construction. You may wish to turn this flag on in order to benchmark style recalculation against other browser engines.

show_debug_borders

True if we should show borders on all layers and tiles for debugging purposes (--show-debug-borders).

show_debug_fragment_borders

True if we should show borders on all fragments for debugging purposes (--show-debug-fragment-borders).

show_debug_parallel_paint

True if we should paint tiles with overlays based on which thread painted them.

show_debug_parallel_layout

True if we should paint borders around flows based on which thread painted them.

paint_flashing

True if we should paint tiles a random color whenever they're repainted. Useful for debugging invalidation.

enable_text_antialiasing

If set with --disable-text-aa, disable antialiasing on fonts. This is primarily useful for reftests where pixel perfect results are required when using fonts such as the Ahem font for layout tests.

enable_canvas_antialiasing

If set with --disable-canvas-aa, disable antialiasing on the HTML canvas element. Like --disable-text-aa, this is useful for reftests where pixel perfect results are required.

trace_layout

True if each step of layout is traced to an external JSON file for debugging purposes. Settings this implies sequential layout and paint.

profile_script_events

Periodically print out on which events script threads spend their processing time.

profile_heartbeats

Enable all heartbeats for profiling.

devtools_port

None to disable devtools or Some with a port number to start a server to listen to remote Firefox devtools connections.

webdriver_port

None to disable WebDriver or Some with a port number to start a server to listen to remote WebDriver commands.

initial_window_size

The initial requested size of the window.

user_agent

An optional string allowing the user agent to be set for testing.

multiprocess

Whether we're running in multiprocess mode.

sandbox

Whether we're running inside the sandbox.

random_pipeline_closure_probability

Probability of randomly closing a pipeline, used for testing the hardening of the constellation.

random_pipeline_closure_seed

The seed for the RNG used to randomly close pipelines, used for testing the hardening of the constellation.

dump_flow_tree

Dumps the flow tree after a layout.

dump_display_list

Dumps the display list after a layout.

dump_display_list_json

Dumps the display list in JSON form after a layout.

dump_layer_tree

Dumps the layer tree when it changes.

relayout_event

Emits notifications when there is a relayout.

disable_share_style_cache

Whether Style Sharing Cache is used

convert_mouse_to_touch

Translate mouse input into touch events.

exit_after_load

True to exit after the page load (-x).

no_native_titlebar

Do not use native titlebar

enable_vsync

Enable vsync in the compositor

use_webrender

True to enable the webrender painting/compositing backend.

webrender_stats

True to show webrender profiling stats on screen.

use_msaa

True if WebRender should use multisample antialiasing.

profile_dir

Directory path for persistent session

render_api

full_backtraces