Struct util::workqueue::WorkQueue

pub struct WorkQueue<QueueData: 'static, WorkData: 'static + Send> {
    workers: Vec<WorkerInfo<QueueData, WorkData>>,
    port: Receiver<SupervisorMsg<QueueData, WorkData>>,
    work_count: usize,
}

A work queue on which units of work can be submitted.

Fields

workers

Information about each of the workers.

port

A port on which deques can be received from the workers.

work_count

The amount of work that has been enqueued.

Methods

impl<QueueData: Sync, WorkData: Send> WorkQueue<QueueData, WorkData>

fn new(thread_name: &'static str, state: ThreadState, thread_count: usize) -> WorkQueue<QueueData, WorkData>

Creates a new work queue and spawns all the threads associated with it.

fn push(&mut self, work_unit: WorkUnit<QueueData, WorkData>)

Enqueues a block into the work queue.

fn run(&mut self, data: &QueueData)

Synchronously runs all the enqueued tasks and waits for them to complete.

fn heap_size_of_tls(&self, f: fn() -> usize) -> Vec<usize>

Synchronously measure memory usage of any thread-local storage.

fn shutdown(&mut self)